Eliminate 'possible loss of data' compiler warning in GC_envfile_getenv (fix of commit 176d5bda1)
* misc.c [GC_READ_ENV_FILE] (GC_envfile_getenv): Change type of namelen
from unsigned to size_t.

+++ misc.c
@@ -719,7 +719,8 @@ GC_API void GC_CALL GC_get_heap_usage_safe(GC_word *pheap_size,
   {
     char *p;
     char *end_of_content;
-    unsigned namelen;
+    size_t namelen;
+
 #   ifndef NO_GETENV
       p = getenv(name); /* try the standard getenv() first */
       if (p != NULL)
